kayunkim/Team Projects/WTD_PJT(2020.08.31~2020.10.08)/Django/requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==2.1.15
django-allauth==0.42.0
django-cors-headers==3.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.10
jwt==1.0.0
oauthlib==3.1.0
pycparser==2.20
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
requests==2.23.0
absl-py==0.10.0
aiohttp==3.6.2
appdirs==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.6.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.6.0
distlib==0.3.1
Django==2.1.15
django-allauth==0.42.0
django-cors-headers==3.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
filelock==3.0.12
gast==0.3.3
google-api-core==1.22.2
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-speech==2.0.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
joblib==0.16.0
JPype1==1.0.2
jwt==1.0.0
Keras-Preprocessing==1.1.2
konlpy==0.5.2
libcst==0.3.11
lxml==4.5.2
Markdown==3.3
multidict==4.7.6
mypy-extensions==0.4.3
mysql-connector-python==8.0.21
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.3
Pillow==7.2.0
proto-plus==1.10.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.9.27
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.50.0
tweepy==3.9.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
